extract custom hooks for top artists and genres
